+++ core/layout/main.php
@@ -167,14 +167,38 @@
 		} else {
 		?>
 		<div class="row">
-			<?php if ($blockleft !== "") :?> <div class="<?php echo $blockleft; ?>" id="contentleft">
-			<?php
-			 echo $this->getData(['page',$this->getData(['page',$this->getUrl(0),'barLeft']),'content']);
-			 ?></div> <?php endif; ?>
-			<div class="<?php echo $content; ?>" id="contentsite"><?php $layout->showContent(); ?></div>
-			<?php if ($blockright !== "") :?> <div class="<?php echo $blockright; ?>" id="contentright">
-			<?php echo $this->getData(['page',$this->getData(['page',$this->getUrl(0),'barRight']),'content']);
-			?></div> <?php endif; ?>
+			<?php 
+				if ($blockleft !== "") :?> 
+				<div class="<?php echo $blockleft; ?>" id="contentLeft">		
+					<?php
+					// Détermine si le bloc a un menu à inclure
+					
+					if ($this->getData(['page',$this->getData(['page',$this->getUrl(0),'barLeft']),'displayMenu']) !== '0') {
+						?> <div id="menuside"><nav> <?php
+						$filter =  ($this->getData(['page',$this->getData(['page',$this->getUrl(0),'barLeft']),'displayMenu']) === '1') ? false : true;
+						$layout->showMenuSide($filter);
+						?></nav></div><?php
+					}					
+					$layout->showBarContentLeft(); ?>
+			 	</div> 
+				<?php endif; ?>
+			<div class="<?php echo $content; ?>
+				" id="contentSite"><?php $layout->showContent(); ?>
+			</div>
+			<?php 
+				if ($blockright !== "") :?> 
+				<div class="<?php echo $blockright; ?>" id="contentRight">
+					<?php
+					// Détermine si le bloc a un menu à inclure
+					if ($this->getData(['page',$this->getData(['page',$this->getUrl(0),'barRight']),'displayMenu']) !== '0') {
+						?> <div id="menuside"><nav> <?php
+						$filter =  ($this->getData(['page',$this->getData(['page',$this->getUrl(0),'barRight']),'displayMenu']) === '1') ? false : true;
+						$layout->showMenuSide($filter);
+						?></nav></div><?php						
+					}
+					$layout->showBarContentRight(); ?>
+				</div>
+				<?php endif; ?>	
 		</div>
 		<?php } ?>
 	</section>
@@ -291,8 +315,7 @@
 	</footer>
 <?php endif; ?>
 <!-- Lien remonter en haut -->
-<div id="backToTop"><?php echo template::ico('chevron-up'); ?></div>
-<?php $layout->showAnalytics(); ?>
+<div id="backToTop"><?php echo template::ico('up'); ?></div>
 <?php $layout->showScript(); ?>
 </body>
 </html>