+++ core/module/theme/view/footer/footer.php
@@ -73,7 +73,7 @@
 					<div class="col3">
 						<?php echo template::select('themeFooterFontSize', $module::$footerFontSizes, [
 							'label' => 'Taille',
-							'help' => 'Proportionnelle à celle définie dans le site',							
+							'help' => 'Proportionnelle à celle définie dans le site.',							
 							'selected' => $this->getData(['theme', 'footer', 'fontSize'])
 						]); ?>
 					</div>